Removes density from rate formulation
I made a mistake yesterday when formulating the sf and cracking rates.
For some reason I took [Tar] * A * exp(...) and divided it by rho. The
thought was that I wanted units of [1/s] and the paper (brown 1998) rates had
units of [kg/(m^3 s)]. But I neglected that in the paper they multiply with
the density so by dividing I jsut get rid of the density already there.

It is also clear just looking at the equation that ``[Tar] * A * exp(...)"
will have the desired units and dividing it by rho messes things up.
This catch has the happy result of making the secondary tar integration
at least become positive.