Converts comparison.py funcitonality into a Class
We need to have a hook and be able to return things based on changes in the
rate constants in order to perform the optimization. The original
comparison file was for jsut comparing the two. It was relatively
straightforward to convert it into a Class that has a integrateRate
function which takes the rate constants as a parameter.