New commit after Masters
I apparently made quite a few changes to this repo and
did not push them to github near the end of my Masters.

I no longer have access to the files on my work computer and
the .git files were not copied to the backups that are now
all I have to work with.
So I cloned the repo (now outdated by about a month) to my
personal computer then deleted everything in the folder
and copied the from my backup (which is the most recent work)
into the newly cloned and deleted repo. Dont tell Linus.